Revert “Add something that might become useful later”
This reverts commit abf213d4e64c558aec17ccc37e63c02a3cbbf0bc.

I am not going to be using this, either.

I was thinking of using Git to implementing pushing and popping notes.
This would involve outputting the additions of a commit only, and
without the `+` prefixes.

Anyway.  It wouldn’t have been an elegant solution, it wouldn’t have
been *efficient*, but I thought it might be fun.  But now I’ve decided
against this approach.