Wrap the script (so far) in a subroutine
Wrap the routine for adding a “jot” in a subroutine.  I intend to call
this subroutine in case there are no arguments to the command; the other
case is when the argument to the command is `pop`, in which case the
latest jot shall be removed from the file and printed.